backend/requirements.txt
fastapi
uvicorn[standard]
ytmusicapi
python-dotenv
pydantic
yt-dlp
